+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed deeply flawed adulthood detection ([#276](https://gitlab.com/Skn0tt/EntE/issues/276))
 - Add further documentation on administration and usage ([#161](https://gitlab.com/Skn0tt/EntE/issues/161))
+- Update to react-router@5.0.0 ([#216](https://gitlab.com/Skn0tt/EntE/issues/216))
 - Made entry creation deadline configurable ([#260](https://gitlab.com/Skn0tt/EntE/issues/260))
 - Fetching entries/slots/users removes potentially outdated data ([#265](https://gitlab.com/Skn0tt/EntE/issues/265))
 - Fixed bug in user creation form where old data triggered validation ([#274](https://gitlab.com/Skn0tt/EntE/issues/274))